fix(core): encode ownerIndex as the uint256 the contract decodes
`SignatureWrapper.ownerIndex` is `uint256` in the struct and `wrapSignature`
declared it `uint8`. The bytes are identical for anything that fits in a byte,
so nothing in use today moves, but viem throws IntegerOutOfRangeError past 255
for an index the contract accepts.

`MultiOwnable` hands out indices from a counter that only goes up and removing
an owner does not reuse or compact them, so the ceiling is on how many owners
the account has ever had. An owner sitting above index 255 could not sign
through this SDK at all.

Adds a vector at index 300, derived with `cast` like the others, which the old
declaration threw on. The three existing vectors are unchanged, which is the
point: widening does not move the bytes.

+++ packages/core/src/account/toJustanAccount.test.ts
@@ -23,7 +23,7 @@ const WRAPPED_SIGNATURE = [
     {
         type: 'tuple',
         components: [
-            { name: 'ownerIndex', type: 'uint8' },
+            { name: 'ownerIndex', type: 'uint256' },
             { name: 'signatureData', type: 'bytes' },
         ],
     },
@@ -595,7 +595,7 @@ describe('toJustanAccount unit tests', () => {
                 // Decoded, not "was called": the wrapping is what the contract
                 // reads, and toJustanAccount.encoding.test.ts pins its shape.
                 const [wrapped] = viem.decodeAbiParameters(WRAPPED_SIGNATURE, signature);
-                expect(wrapped.ownerIndex).toBe(0);
+                expect(wrapped.ownerIndex).toBe(0n);
                 expect(wrapped.signatureData).toBe(MOCK_SIGNATURE);
             });
 
@@ -682,7 +682,7 @@ describe('toJustanAccount unit tests', () => {
                 // out above, wrapped by viem's own helper.
                 const { wrapTypedDataSignature } = await import('viem/experimental/erc7739');
                 const inner = viem.encodeAbiParameters(WRAPPED_SIGNATURE, [
-                    { ownerIndex: 0, signatureData: MOCK_SIGNATURE },
+                    { ownerIndex: 0n, signatureData: MOCK_SIGNATURE },
                 ]);
                 expect(signature).toBe(wrapTypedDataSignature({ ...MOCK_TYPED_DATA, signature: inner }));
             });
@@ -781,7 +781,7 @@ describe('toJustanAccount unit tests', () => {
                 } as any);
 
                 const [wrapped] = viem.decodeAbiParameters(WRAPPED_SIGNATURE, signature);
-                expect(wrapped.ownerIndex).toBe(0);
+                expect(wrapped.ownerIndex).toBe(0n);
                 expect(wrapped.signatureData).toBe(MOCK_SIGNATURE);
             });
